doom: Add -pistolstart command line option
When the user includes -pistolstart as a command line option, it causes
there to be an automatic "pistol start" when advancing from one level to
the next. Specifically, the player's health is reset to 100, their armor
to 0 and their inventory is reduced to the following: pistol, fists and
50 bullets. This option has no effect when recording a demo, playing
back a demo or when starting a network game.