STbar weapons
No Hell Staff at the moment. It shares a lot of bright pixels on red energy and  on staff itself, which makes bmap looking ugly. Without that bright pixels there will be notable "holes" on red energy... Probably better just leave it as is.

Same problem with one bright pixel on  Ethereal Crossbow, but at least it is possible to avoid it.