strife: Fix popup screen resizing
When showing a popup, decrease the screen size to show the status bar only if the original fullscreen HUD is being used. Then, when the popup closes, increase the screen size to show the original fullscreen HUD again. This preserves the Crispy HUD view and any screen size changes made while the popup was active.