Allow netplay with sideloaded WADs
NET_SV_ParseGameStart checks that episode and map are in the correct
range. This prevented the episode 5 version of Sigil, as well as
sideloaded NRFTL and Master Levels from working in netgames.

This is fixed by changing the number of episodes in the valid_modes
table.  As a side effect, setup program will now allow warping to
E5Mx for the Ultimate Doom IWAD. Outside of Heretic, the valid_modes
table is only used for these two things (netgame validation and setup
warp UI).

Note that if you try to start a netgame with a non-existent map, the
client will simply exit with a W_GetNumForName error.