man/*.template: Break paragraph after copyright
It looks better, and is more conventional, to have a paragraph break
between a list of copyright notices and license texts.

See, for example:

https://spdx.org/licenses/BSD-3-Clause.html
https://spdx.org/licenses/MIT.html
https://spdx.org/licenses/Artistic-2.0.html

Longer licenses, like the GNU GPL and Apache 2, tend to separate the
full texts of their licenses into separate files or sections of a
document, but where the copyright notices are placed, they too are
followed by a paragraph break.

See, for example, the "Standard License Header", the final sections of
both those licenses

https://spdx.org/licenses/GPL-3.0-or-later.html
https://spdx.org/licenses/Apache-2.0.html

Finally, I observe that having the copyright and license notice in the
rendered text of a man page, while sometimes seen, is not common
practice.  Putting this information in *roff comments in the man(7)
source document (in *roff comments) suffices.  It is installed on
systems (if sometimes in compressed form), so there is no risk of the
notices getting separated from the text to which they apply--not by mere
neglect or accident.  Only deliberate action can efface the notices.