man/*.template: Italicize commercial game names
...when introducing them in the lead paragraph of the "DESCRIPTION"
section.  MLA style is to italicize such a thing, because it is "a work
that is complete in and of itself, like a book or movie or painting".

https://universitypark-lonestar.libanswers.com/faq/278770

This is the only instance where such styling seems important.  The game
engines and artistic assets are severable (and severed in practice with
strongly differentiated licensing) and dealt with independently by
source port project developers and users.