man/docgen (manpage_output): Add special cases
...to handle command-line options with complex argument structures
(where "complex" means "anything but the most trivial case").  This
causes them to format idiomatically, with parameters in italics,
literals in bold, and "synopsis language", like brackets denoting
optional parameters, in roman.

This approach is necessary for idiomatic man page rendering because in
the language interpreted by this script, there is no way to mark up such
distinctions in code comments.  Such a mini-language would have to be
designed, and that way lies perlpod(1) and similar efforts.