Fix for sliding doors being able to be manually closed in a half-state (looks closed, actually completely open)
If a player stands with their nose crossing into the sliding door sector and manually closes the door, the code doesn't do the same checks it does in earlier code for the automatic closing of a sliding door.

This allows one to manually close the door but ALSO interrupt the actual ceiling movement and application of line blocking flags. This leaves a sliding door which animates closed but remains completely open, passable, and doesn't block sight, sound, etc.

Basically, I duplicated the checks from the automatic closing routine so a manual close checks for a "crushed" actor (ceiling blocked) AND (and here is where my last deleted pull request comes in, as I misunderstood sec->thinglist, thinking it also contained the player) it checks for a mobj in the way in general, so it won't close on items.

Remove the " && sec->thinglist == NULL" if that last behavior is not vanilla. I added it because I *think* vanilla didn't close these doors on items, as that very check is also in the auto-close routine.